pyproject.toml
[project]
name = "fluid-simu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "glfw>=2.10.0",
    "numpy>=2.4.0",
    "pyopengl>=3.1.10",
]
